Split external_calls functionality
