+++ src/lib.rs
@@ -40,6 +40,7 @@ use self::utils::*;
 
 pub mod capacity_estimator;
 pub mod external_calls;
+pub mod run;
 pub mod toolset;
 // pub mod circuit_limit_estimator;
